requirements.txt
streamlit
yfinance
pandas
plotly
pytz
numpy
google-genai
python-dotenv